SchemaData::enumeration
