+++ tiledb/api/src/array/schema/mod.rs
@@ -24,6 +24,7 @@ use crate::Datatype;
 use crate::Result as TileDBResult;
 
 pub use self::evolution::{Builder as SchemaEvolutionBuilder, SchemaEvolution};
+pub use tiledb_common::array::schema::EnumerationKey;
 pub use tiledb_common::array::{ArrayType, CellValNum};
 
 /// Wrapper for the CAPI handle.
@@ -479,26 +480,6 @@ impl Debug for Schema {
                 write!(f, "<Schema @ {:?}: serialization error: {}>", ptr, e)
             }
         }
-    }
-}
-
-/// A key used to look up an enumeration from a [Schema].
-pub enum EnumerationKey<'a> {
-    /// Identifies an enumeration by its name.
-    EnumerationName(&'a str),
-    /// Identifies an enumeration by the name of an attribute which refers to it.
-    AttributeName(&'a str),
-}
-
-impl<'a> EnumerationKey<'a> {
-    /// Returns [Self::EnumerationName] for any argument which can be used as a [str].
-    pub fn by_name(s: &'a impl AsRef<str>) -> Self {
-        Self::EnumerationName(s.as_ref())
-    }
-
-    /// Returns [Self::AttributeName] for any argument which can be used as a [str].
-    pub fn by_attribute_name(s: &'a impl AsRef<str>) -> Self {
-        Self::AttributeName(s.as_ref())
     }
 }
